+++ coordinator/src/main.rs
@@ -31,9 +31,7 @@ use tokio::{
   time::sleep,
 };
 
-use ::tributary::{
-  ProvidedError, TransactionKind, TransactionError, TransactionTrait, Block, Tributary,
-};
+use ::tributary::{ProvidedError, TransactionKind, TransactionTrait, Block, Tributary};
 
 mod tributary;
 use crate::tributary::{
@@ -126,48 +124,6 @@ async fn add_tributary<D: Db, Pro: Processors, P: P2p>(
     .send(TributaryEvent::NewTributary(ActiveTributary { spec, tributary: Arc::new(tributary) }))
     .map_err(|_| "all ActiveTributary recipients closed")
     .unwrap();
-}
-
-async fn publish_signed_transaction<D: Db, P: P2p>(
-  txn: &mut D::Transaction<'_>,
-  tributary: &Tributary<D, Transaction, P>,
-  tx: Transaction,
-) {
-  log::debug!("publishing transaction {}", hex::encode(tx.hash()));
-
-  let (order, signer) = if let TransactionKind::Signed(order, signed) = tx.kind() {
-    let signer = signed.signer;
-
-    // Safe as we should deterministically create transactions, meaning if this is already on-disk,
-    // it's what we're saving now
-    SignedTransactionDb::set(txn, &order, signed.nonce, &tx.serialize());
-
-    (order, signer)
-  } else {
-    panic!("non-signed transaction passed to publish_signed_transaction");
-  };
-
-  // If we're trying to publish 5, when the last transaction published was 3, this will delay
-  // publication until the point in time we publish 4
-  while let Some(tx) = SignedTransactionDb::take_signed_transaction(
-    txn,
-    &order,
-    tributary
-      .next_nonce(&signer, &order)
-      .await
-      .expect("we don't have a nonce, meaning we aren't a participant on this tributary"),
-  ) {
-    // We need to return a proper error here to enable that, due to a race condition around
-    // multiple publications
-    match tributary.add_transaction(tx.clone()).await {
-      Ok(_) => {}
-      // Some asynchonicity if InvalidNonce, assumed safe to deterministic nonces
-      Err(TransactionError::InvalidNonce) => {
-        log::warn!("publishing TX {tx:?} returned InvalidNonce. was it already added?")
-      }
-      Err(e) => panic!("created an invalid transaction: {e:?}"),
-    }
-  }
 }
 
 // TODO: Find a better pattern for this
@@ -726,7 +682,7 @@ async fn handle_processor_message<D: Db, P: P2p>(
         }
         TransactionKind::Signed(_, _) => {
           tx.sign(&mut OsRng, genesis, key);
-          publish_signed_transaction(&mut txn, tributary, tx).await;
+          tributary::publish_signed_transaction(&mut txn, tributary, tx).await;
         }
       }
     }
@@ -1141,7 +1097,7 @@ pub async fn run<D: Db, Pro: Processors, P: P2p>(
           // TODO: Should this not take a txn accordingly? It's best practice to take a txn, yet
           // taking a txn fails to declare its achieved independence
           let mut txn = raw_db.txn();
-          publish_signed_transaction(&mut txn, tributary, tx).await;
+          tributary::publish_signed_transaction(&mut txn, tributary, tx).await;
           txn.commit();
           break;
         }
