+++ coordinator/src/tributary/mod.rs
@@ -24,9 +24,9 @@ use processor_messages::coordinator::SubstrateSignableId;
 use serai_client::{primitives::PublicKey, validator_sets::primitives::ValidatorSet};
 
 use tributary::{
-  ReadWrite,
+  TRANSACTION_SIZE_LIMIT, ReadWrite,
   transaction::{Signed, TransactionError, TransactionKind, Transaction as TransactionTrait},
-  TRANSACTION_SIZE_LIMIT,
+  Tributary,
 };
 
 mod db;
@@ -809,6 +809,48 @@ impl Transaction {
       Ristretto::hash_to_F(b"SignCompleted signature", &transcript.challenge(b"challenge"))
     } else {
       panic!("sign_completed_challenge called on transaction which wasn't SignCompleted")
+    }
+  }
+}
+
+pub async fn publish_signed_transaction<D: Db, P: crate::P2p>(
+  txn: &mut D::Transaction<'_>,
+  tributary: &Tributary<D, Transaction, P>,
+  tx: Transaction,
+) {
+  log::debug!("publishing transaction {}", hex::encode(tx.hash()));
+
+  let (order, signer) = if let TransactionKind::Signed(order, signed) = tx.kind() {
+    let signer = signed.signer;
+
+    // Safe as we should deterministically create transactions, meaning if this is already on-disk,
+    // it's what we're saving now
+    SignedTransactionDb::set(txn, &order, signed.nonce, &tx.serialize());
+
+    (order, signer)
+  } else {
+    panic!("non-signed transaction passed to publish_signed_transaction");
+  };
+
+  // If we're trying to publish 5, when the last transaction published was 3, this will delay
+  // publication until the point in time we publish 4
+  while let Some(tx) = SignedTransactionDb::take_signed_transaction(
+    txn,
+    &order,
+    tributary
+      .next_nonce(&signer, &order)
+      .await
+      .expect("we don't have a nonce, meaning we aren't a participant on this tributary"),
+  ) {
+    // We need to return a proper error here to enable that, due to a race condition around
+    // multiple publications
+    match tributary.add_transaction(tx.clone()).await {
+      Ok(_) => {}
+      // Some asynchonicity if InvalidNonce, assumed safe to deterministic nonces
+      Err(TransactionError::InvalidNonce) => {
+        log::warn!("publishing TX {tx:?} returned InvalidNonce. was it already added?")
+      }
+      Err(e) => panic!("created an invalid transaction: {e:?}"),
     }
   }
 }
